Added -copts and -bopts, bug fixes
Fixed bug in which having multiple HSPs for the same subject BLAST hit resulted in duplicated sequences in the final SVG output.

Added the -copts and -bopts options, which allow the user to specify their own clustalo and blastp options (with some exceptions). -copts and -bopts must be in the quoted, bracketed format specified by the help descriptions.