Bug fixes and preliminary features
Fixed a bug in which not specifying -bopts sent an empty option to blastp.

retrieve_annotations.py now generates an all.metadata file that contains EC numbers and descriptions (retrieved from UniProt entries) for each subject hit for a particular query. There are plans to add PDB IDs to this file as well. For now, this column is saved as "placeholder".